change API

+++ src/main/java/co/edu/unbosque/ciclo3back/dao/clientesDAO.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.ciclo3back.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import co.edu.unbosque.ciclo3back.model.clientes;
 
-public interface clientesDAO extends JpaRepository<clientes, Integer> {
+public interface clientesDAO extends JpaRepository<clientes, Long> {
 
 }
