+++ src/main/java/co/edu/unbosque/ciclo3back/dao/productosDAO.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.ciclo3back.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import co.edu.unbosque.ciclo3back.model.productos;
 
-public interface productosDAO extends JpaRepository<productos, Integer> {
+public interface productosDAO extends JpaRepository<productos, Long> {
 
 }
