+++ src/main/java/co/edu/unbosque/ciclo3back/dao/proveedoresDAO.java
@@ -3,6 +3,6 @@ package co.edu.unbosque.ciclo3back.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import co.edu.unbosque.ciclo3back.model.proveedores;
 
-public interface proveedoresDAO extends JpaRepository<proveedores, Integer> {
+public interface proveedoresDAO extends JpaRepository<proveedores, Long> {
 
 }
